Convert storage size before considering resizing
